@@ -5,14 +5,11 @@ class Responsive extends StatelessWidget {
   final Widget desktop;
 
   const Responsive({
-    Key key,
-    @required this.mobile,
-    @required this.desktop,
+    Key? key,
+    required this.mobile,
+    required this.desktop,
   }) : super(key: key);
 
-// This size work fine on my design, maybe you need some customization depends on your design
-
-  // This isMobile, isTablet, isDesktop helep us later
   static bool isMobile(BuildContext context) =>
       MediaQuery.of(context).size.width < 650;
 
@@ -22,7 +19,6 @@ class Responsive extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return LayoutBuilder(
-      // If our width is more than 1100 then we consider it a desktop
       builder: (context, constraints) {
         if (constraints.maxWidth >= 650) {
           return desktop;
